Split massive server.views up.
Newly divided into:
1. Settings UI views.
1. "regular" UI views.
1. Non UI views (checkins, csv, etc).

I then cleaned out all of the now-unused imports for each of these.

Now vim doesn't freak out when I open views.py ;)
